reorg optimizer fixtures for reuse
